pyproject.toml
[project]
name = "nyc-taxi-analytics"
version = "0.1.0"
description = "Real-time analytics platform for NYC TLC taxi trip data using GCP managed services"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [{ name = "NYC Taxi Analytics Team" }]

dependencies = [
    # Core
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",

    # Data processing
    "pandas>=2.1.0",
    "pyarrow>=14.0.0",

    # GCP
    "google-cloud-pubsub>=2.18.0",
    "google-cloud-storage>=2.14.0",
    "google-cloud-bigquery>=3.14.0",

    # Beam/Dataflow
    "apache-beam[gcp]>=2.52.0",

    # API
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "httpx>=0.26.0",

    # Airflow (for local development)
    "apache-airflow>=2.8.0",
    "apache-airflow-providers-google>=10.13.0",

    # CLI
    "typer>=0.9.0",
    "rich>=13.7.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "pytest-asyncio>=0.23.0",
    "black>=24.1.0",
    "ruff>=0.1.14",
    "mypy>=1.8.0",
    "pre-commit>=3.6.0",
    "types-requests>=2.31.0",
]

[project.scripts]
taxi-producer = "src.producer.main:app"
taxi-pipeline = "src.pipeline.main:main"
taxi-api = "src.api.main:run"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.black]
line-length = 100
target-version = ["py311"]
include = '\.pyi?$'
exclude = '''
/(
    \.git
    | \.hg
    | \.mypy_cache
    | \.tox
    | \.venv
    | _build
    | buck-out
    | build
    | dist
)/
'''

[tool.ruff]
line-length = 100
target-version = "py311"
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # Pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
]
ignore = [
    "E501",  # line too long (handled by black)
    "B008",  # do not perform function calls in argument defaults
]

[tool.ruff.isort]
known-first-party = ["src"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_ignores = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
strict_optional = true
plugins = ["pydantic.mypy"]

[[tool.mypy.overrides]]
module = [
    "apache_beam.*",
    "google.cloud.*",
    "airflow.*",
]
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
addopts = "-v --cov=src --cov-report=term-missing"
asyncio_mode = "auto"
